billing/api-gateway: rename internal package to pkg.
