+++ systems/billing/api-gateway/pkg/client/pdf.go
