+++ systems/billing/api-gateway/pkg/global.go
@@ -1,4 +1,4 @@
-package internal
+package pkg
 
 const ServiceName = "gateway"
 const SystemName = "billing"
